requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.2
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
proxy-tools==0.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.16.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.15
pywebview==5.3.2
QtPy==2.4.2
typing_extensions==4.12.2
Werkzeug==3.1.3
